Updated long long to int64_t

+++ asm/include/asm/assembler.h
@@ -19,8 +19,8 @@ public:
    * then we encode an REX byte and do not encode one otherwise.
    */
   void append(instruction inst, size operandSize, size addressSize, reg_type regType,
-  addressing_mode addrMode, reg rmReg, long rmDisp, 
-  reg rReg, bool isImm, long imm, bool rmIsDest, bool uniformReg,
+  addressing_mode addrMode, reg rmReg, long long rmDisp, 
+  reg rReg, bool isImm, long long imm, bool rmIsDest, bool uniformReg,
   std::vector<unsigned char>& buf);
   std::vector<unsigned char> assemble(std::string str);
 };
